ESLint: remove redundant disable of vars-on-top rule
No longer needed now that we use eslint-config-wikimedia
version 0.20.0, which no longer includes that rule.

If the rule was set to a warning instead of disabled entirely,
I still count that as redundant because the purpose of that
is generally to make the tests not fail (by default they only
fail if a rule triggers an error) rather than to intentionally
opt-in to the rule as a warning.

Bug: T287796
Change-Id: I1f3c6c7be6d5ced366e92a1b11ddc77211fc3a20